Allow modifying the show clause.
--There are situations in which customers want to be able to use
dimension fields for internal processing only, without exposing them to
users. However, currently that is extremely difficult, because no extension point
exists for cleanly restricting which dimension fields a user is or is
not allowed to see.

Therefore, we make two changes: First, we add a wither to
`DataApiRequest` that allows us to replace the dimension fields to be
shown.

Second, the `DimensionsApiRequest` now generates the dimension fields to
expose in much the same way that the `DataApiRequest` does (defaulting
to all the dimension fields for every dimension requested), though it
does not attempt to give the user the ability to restrict the fields
seen. We also add a wither to `DimensionsApiRequest` to allow mappers to
overwrite the fields, and the `DimensionsServlet` extracts the dimension
fields to display from the `DimensionsApiRequest` rather than from the
dimension being displayed.

While the longterm solution is to add some sort of infrastructure
around dimension fields that allow customers to tag certain fields, and
then restrict their usage based on those tags, this provides a very
cheap, minimally invasive way of allowing customers to customize which fields
are or are not displayed via a `RequestMapper`.